oop_version/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
coverage==7.13.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
